Split jms into seperate modules for maintainability. Use __all__ metavars to avoid namespace clutter.
